dead16: remove dead functions from pkg/cli (12 functions)
