+++ pkg/cli/commands_test.go
@@ -230,113 +230,6 @@ func TestStatusWorkflows(t *testing.T) {
 	}
 }
 
-func TestEnableWorkflows(t *testing.T) {
-	err := EnableWorkflows("test-pattern")
-
-	// Should now return an error since no workflows can be found to enable
-	if err == nil {
-		t.Errorf("EnableWorkflows should return error when no workflows found to enable, got nil")
-	}
-
-	// The error should indicate workflows couldn't be found
-	if !strings.Contains(err.Error(), "workflow") {
-		t.Errorf("EnableWorkflows error should mention workflows, got: %v", err)
-	}
-}
-
-func TestEnableWorkflowsFailureScenarios(t *testing.T) {
-	tests := []struct {
-		name          string
-		pattern       string
-		expectError   bool
-		errorContains string
-		description   string
-	}{
-		{
-			name:          "empty pattern",
-			pattern:       "",
-			expectError:   true,
-			errorContains: "workflow",
-			description:   "Should error when no workflows found to enable",
-		},
-		{
-			name:          "nonexistent pattern",
-			pattern:       "nonexistent-workflow-pattern",
-			expectError:   true,
-			errorContains: "workflow",
-			description:   "Should error when pattern matches no workflows",
-		},
-		{
-			name:          "wildcard pattern",
-			pattern:       "xyz-*",
-			expectError:   true,
-			errorContains: "workflow",
-			description:   "Should error when wildcard pattern matches nothing",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := EnableWorkflows(tt.pattern)
-
-			if tt.expectError {
-				if err == nil {
-					t.Errorf("%s: expected error but got nil", tt.description)
-				} else if !strings.Contains(err.Error(), tt.errorContains) {
-					t.Errorf("%s: expected error containing '%s', got: %v", tt.description, tt.errorContains, err)
-				}
-			} else {
-				if err != nil {
-					t.Errorf("%s: unexpected error: %v", tt.description, err)
-				}
-			}
-		})
-	}
-}
-
-func TestDisableWorkflows(t *testing.T) {
-	err := DisableWorkflows("test-pattern")
-
-	// Disable should now also return an error when no workflows found
-	if err == nil {
-		t.Errorf("DisableWorkflows should return error when no workflows found to disable, got nil")
-	}
-
-	// The error should indicate workflows couldn't be found
-	if !strings.Contains(err.Error(), "workflow") {
-		t.Errorf("DisableWorkflows error should mention workflows, got: %v", err)
-	}
-}
-
-func TestDisableWorkflowsFailureScenarios(t *testing.T) {
-	// Test that disable now also errors on failure scenarios
-	tests := []struct {
-		name        string
-		pattern     string
-		description string
-	}{
-		{
-			name:        "empty pattern",
-			pattern:     "",
-			description: "Should error when no workflows found to disable",
-		},
-		{
-			name:        "nonexistent pattern",
-			pattern:     "nonexistent-workflow-pattern",
-			description: "Should error when pattern matches no workflows",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := DisableWorkflows(tt.pattern)
-			if err == nil {
-				t.Errorf("%s: DisableWorkflows should now error, got nil", tt.description)
-			}
-		})
-	}
-}
-
 func TestRunWorkflowOnGitHub(t *testing.T) {
 	// Test with empty workflow name
 	err := RunWorkflowOnGitHub(context.Background(), "", RunOptions{})
@@ -479,8 +372,6 @@ Test workflow for command existence.`
 		}, false, "CompileWorkflows"},
 		{func() error { return RemoveWorkflows("nonexistent", false) }, false, "RemoveWorkflows"},                    // Should handle missing directory gracefully
 		{func() error { return StatusWorkflows("nonexistent", false, false, "", "", "") }, false, "StatusWorkflows"}, // Should handle missing directory gracefully
-		{func() error { return EnableWorkflows("nonexistent") }, true, "EnableWorkflows"},                            // Should now error when no workflows found to enable
-		{func() error { return DisableWorkflows("nonexistent") }, true, "DisableWorkflows"},                          // Should now also error when no workflows found to disable
 		{func() error {
 			return RunWorkflowOnGitHub(context.Background(), "", RunOptions{})
 		}, true, "RunWorkflowOnGitHub"}, // Should error with empty workflow name
