+++ pkg/cli/docker_images.go
@@ -225,39 +225,3 @@ func CheckAndPrepareDockerImages(ctx context.Context, useZizmor, usePoutine, use
 
 	return nil
 }
-
-// isDockerAvailable checks if Docker is available on the system
-func isDockerAvailable() bool {
-	cmd := exec.Command("docker", "version")
-	cmd.Stdout = nil
-	cmd.Stderr = nil
-	return cmd.Run() == nil
-}
-
-// ResetDockerPullState resets the internal pull state (for testing)
-func ResetDockerPullState() {
-	pullState.mu.Lock()
-	defer pullState.mu.Unlock()
-	pullState.downloading = make(map[string]bool)
-	pullState.mockAvailable = make(map[string]bool)
-	pullState.mockAvailableInUse = false
-}
-
-// SetDockerImageDownloading sets the downloading state for an image (for testing)
-func SetDockerImageDownloading(image string, downloading bool) {
-	pullState.mu.Lock()
-	defer pullState.mu.Unlock()
-	if downloading {
-		pullState.downloading[image] = true
-	} else {
-		delete(pullState.downloading, image)
-	}
-}
-
-// SetMockImageAvailable sets the mock availability for an image (for testing)
-func SetMockImageAvailable(image string, available bool) {
-	pullState.mu.Lock()
-	defer pullState.mu.Unlock()
-	pullState.mockAvailableInUse = true
-	pullState.mockAvailable[image] = available
-}
